requirements.txt
python-telegram-bot==22.5
sqlalchemy>=2.0.36
psycopg[binary]
python-dotenv==1.0.0
